#528 RTU gateway configurator cleanup
